Import the beta image
The beta deployment-tools image has been verified to work for
cloud-infra and i'm working through cloud-ops pipelines as well. Thus
far this has been a drop in replacement for the most part. I had to
disable the default entrypoint of /bin/bash due to how jenkins exec's
the container.

Otherwise, this is a fantastic little upgrade for deployment-tools that
i've been running locally for about 2 months without major issue.

It's time to push this to deployment-tools proper and adopt it.